Update index.html
Isako add kiya-
'You can use it like an app. Just open it in the Chrome browser, click on the three dots at the top of the browser, and select "Add to Home screen". Doing this will place the application on your device’s Home screen, and you can use it just like an app!'